Create gpt-commit-summarizer.yml
This workflow file tells GitHub to run the action whenever a new Commit is pushed on the repo or a pr is merged. That's it! You're now ready to use the commit-blog generator action in your repository. Each time a pull request is merged or a commit is pushed, the action will automatically publish a blog on hashnnode that will contain the summary of the changes made in commit.